requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
deprecation==2.1.0
    # via python-keycloak
django==4.2.1
    # via
    #   -r requirements.in
    #   django-filter
    #   django-tenants
    #   djangorestframework
django-environ==0.10.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-tenants==3.5.0
    # via -r requirements.in
django-verify-email==2.0.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
markdown==3.4.3
    # via -r requirements.in
packaging==23.1
    # via deprecation
pillow==9.5.0
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
python-jose==3.3.0
    # via python-keycloak
python-keycloak==3.3.0
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
requests==2.31.0
    # via
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
urllib3==2.0.4
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
